Refactored Files, Added material theming
