+++ lib/DataClasses/Counter.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class Counter with ChangeNotifier {
+  int _counter;
+
+  Counter(this._counter);
+  get counter => this._counter;
+  set counter(int counter) => this._counter = counter;
+
+  void increment() {
+    _counter++;
+    notifyListeners();
+  }
+
+  void decrement() {
+    _counter--;
+    notifyListeners();
+  }
+}
