+++ lib/main.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:provider/provider.dart';
-import 'package:society_app_signup/Components/IconLogo.dart';
+import 'package:society_app_signup/Components/Home.dart';
+import 'package:society_app_signup/DataClasses/Counter.dart';
+import 'package:society_app_signup/Styling/AppTheme.dart';
 
 void main() => runApp(MyApp());
 
@@ -11,9 +12,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
+      theme: appTheme(),
       home: ChangeNotifierProvider<Counter>(
         builder: (_) => Counter(0),
         child: Consumer<Counter>(
@@ -23,75 +22,5 @@ class MyApp extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    final _counter = Provider.of<Counter>(context);
-
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            IconLogo(),
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '${_counter.counter}',
-              style: Theme.of(context).textTheme.display1,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: <Widget>[
-          FloatingActionButton(
-            onPressed: () =>
-              Provider.of<Counter>(context, listen: false).increment(),
-            tooltip: 'Increment',
-            child: Icon(Icons.add),
-          ),
-          SizedBox(height: 10),
-          FloatingActionButton(
-            onPressed: () =>
-              Provider.of<Counter>(context, listen: false).decrement(),
-            tooltip: 'Increment',
-            child: Icon(Icons.remove),
-          )
-        ],
-      ), // This trailing comma makes auto-formatting nicer for build methods.
-    );
-  }
-}
-
-class Counter with ChangeNotifier {
-  int _counter;
-
-  Counter(this._counter);
-  get counter => this._counter;
-  set counter(int counter) => this._counter = counter;
-
-  void increment() {
-    _counter++;
-    notifyListeners();
-  }
-
-  void decrement() {
-    _counter--;
-    notifyListeners();
   }
 }
